Move RC to be command-specific
The RC file is passed into each command and validated individually
instead of being validated at the top level. This allows each
command to define for itself what properties are needed.

Fixes #832.
