+++ packages/brookjs-cli/src/commands/BuildCommand/initialState.ts
@@ -1,8 +1,7 @@
 import { Arguments } from 'yargs';
 import webpack from 'webpack';
-import { Maybe } from 'brookjs-types';
-import { RCResult } from '../../RC';
 import { State } from './types';
+import { RC } from './RC';
 
 const getEnv = (env: unknown): Required<webpack.Configuration>['mode'] => {
   if (
@@ -17,13 +16,13 @@ const getEnv = (env: unknown): Required<webpack.Configuration>['mode'] => {
 
 const initialState = (
   args: Arguments,
-  { rc, cwd }: { rc: Maybe<RCResult>; cwd: string }
+  { rc, cwd }: { rc: unknown; cwd: string }
 ): State => ({
   watch: typeof args.watch === 'boolean' ? args.watch : false,
   building: true,
   results: null,
   env: getEnv(args.env),
-  rc: rc ?? null,
+  rc: RC.decode(rc).getOrElse({}) as RC,
   cwd
 });
